+++ packages/materials/src/ui/index.ts
@@ -24,6 +24,8 @@ import list from './list';
 import queryForm from './query-form';
 import tabs from './tabs';
 import picker from './picker';
+import importButton from './import-button';
+import dialogGrid from './dialog-grid';
 
 const name = '@vtj/ui';
 const components: MaterialDescription[] = [
@@ -45,7 +47,9 @@ const components: MaterialDescription[] = [
   list,
   queryForm,
   tabs,
-  picker
+  picker,
+  importButton,
+  dialogGrid
 ].flat();
 
 const categories: MaterialCategory[] = [