+++ packages/ui/src/components/import-button/index.ts
@@ -1,0 +1,3 @@
+import XImportButton from './ImportButton.vue';
+export { XImportButton };
+export * from './types';