+++ packages/ui/src/components/index.ts
@@ -24,3 +24,5 @@ export * from './attachment';
 export * from './query-form';
 export * from './grid';
 export * from './picker';
+export * from './import-button';
+export * from './dialog-grid';