+++ apps/app/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.77](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.8.76...vtj-project-app@0.8.77) (2024-07-09)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.8.76](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.8.75...vtj-project-app@0.8.76) (2024-07-08)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.8.75](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.8.74...vtj-project-app@0.8.75) (2024-07-08)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.8.74](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.8.73...vtj-project-app@0.8.74) (2024-07-06)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.8.73](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.8.72...vtj-project-app@0.8.73) (2024-07-06)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.8.72](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.8.71...vtj-project-app@0.8.72) (2024-07-06)
+
+
+### Features
+
+* ✨ 支持项目部署二级目录配置 ([200e87b](https://gitee.com/newgateway/vtj/commits/200e87bea6ed66776d367164e2c84584e3cd22a1))
+
+
+
+
+
 ## [0.8.71](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.8.70...vtj-project-app@0.8.71) (2024-07-05)
 
 **Note:** Version bump only for package vtj-project-app